requirements.txt
yt-dlp
openai-whisper
transformers
torch
sentencepiece
sacremoses
jieba  # 用于中文分词
protobuf==3.20.0  # 使用稳定版本
openai  # 添加 OpenAI 依赖 